split helper.go
