+++ pkg/provider/cloud/openstack/helper.go
@@ -19,7 +19,6 @@ package openstack
 import (
 	"errors"
 	"fmt"
-	"net/http"
 	"strings"
 
 	"github.com/gophercloud/gophercloud"
@@ -29,21 +28,7 @@ import (
 	osprojects "github.com/gophercloud/gophercloud/openstack/identity/v3/projects"
 	ostokens "github.com/gophercloud/gophercloud/openstack/identity/v3/tokens"
 	osusers "github.com/gophercloud/gophercloud/openstack/identity/v3/users"
-	osextnetwork "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/external"
-	osrouters "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/layer3/routers"
-	ossecuritygroups "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/groups"
-	ossecuritygrouprules "github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/rules"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/subnetpools"
-	osnetworks "github.com/gophercloud/gophercloud/openstack/networking/v2/networks"
-	osports "github.com/gophercloud/gophercloud/openstack/networking/v2/ports"
-	ossubnets "github.com/gophercloud/gophercloud/openstack/networking/v2/subnets"
 	"github.com/gophercloud/gophercloud/pagination"
-
-	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
-	"k8c.io/kubermatic/v2/pkg/provider"
-
-	"k8s.io/utils/net"
-	"k8s.io/utils/ptr"
 )
 
 const (
@@ -55,546 +40,6 @@ const (
 
 	resourceNamePrefix = "kubernetes-"
 )
-
-func getRouterByName(netClient *gophercloud.ServiceClient, name string) (*osrouters.Router, error) {
-	routers, err := listRouters(netClient, osrouters.ListOpts{Name: name})
-	if err != nil {
-		return nil, err
-	}
-	switch len(routers) {
-	case 1:
-		return &routers[0], nil
-	case 0:
-		return nil, fmt.Errorf("router with name '%s' not found", name)
-	default:
-		return nil, fmt.Errorf("found %d routers for name '%s', expected exactly one", len(routers), name)
-	}
-}
-
-func getRouterByID(netClient *gophercloud.ServiceClient, id string) (*osrouters.Router, error) {
-	routers, err := listRouters(netClient, osrouters.ListOpts{ID: id})
-	if err != nil {
-		return nil, err
-	}
-	switch len(routers) {
-	case 1:
-		return &routers[0], nil
-	case 0:
-		return nil, fmt.Errorf("router with ID '%s' not found", id)
-	default:
-		return nil, fmt.Errorf("found %d routers for ID '%s', expected exactly one", len(routers), id)
-	}
-}
-
-func listRouters(netClient *gophercloud.ServiceClient, listOpts osrouters.ListOpts) ([]osrouters.Router, error) {
-	allPages, err := osrouters.List(netClient, listOpts).AllPages()
-	if err != nil {
-		return nil, err
-	}
-
-	allRouters, err := osrouters.ExtractRouters(allPages)
-	if err != nil {
-		return nil, err
-	}
-	return allRouters, nil
-}
-
-func getSecurityGroups(netClient *gophercloud.ServiceClient, opts ossecuritygroups.ListOpts) ([]ossecuritygroups.SecGroup, error) {
-	page, err := ossecuritygroups.List(netClient, opts).AllPages()
-	if err != nil {
-		return nil, fmt.Errorf("failed to list security groups: %w", err)
-	}
-	secGroups, err := ossecuritygroups.ExtractGroups(page)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract security groups: %w", err)
-	}
-	return secGroups, nil
-}
-
-// NetworkWithExternalExt is a struct that implements all networks.
-type NetworkWithExternalExt struct {
-	osnetworks.Network
-	osextnetwork.NetworkExternalExt
-}
-
-func getAllNetworks(netClient *gophercloud.ServiceClient, opts osnetworks.ListOpts) ([]NetworkWithExternalExt, error) {
-	var allNetworks []NetworkWithExternalExt
-	allPages, err := osnetworks.List(netClient, opts).AllPages()
-	if err != nil {
-		return nil, err
-	}
-
-	if err = osnetworks.ExtractNetworksInto(allPages, &allNetworks); err != nil {
-		return nil, err
-	}
-
-	return allNetworks, nil
-}
-
-func getNetworkByName(netClient *gophercloud.ServiceClient, name string, isExternal bool) (*NetworkWithExternalExt, error) {
-	existingNetworks, err := getAllNetworks(netClient, osnetworks.ListOpts{Name: name})
-	if err != nil {
-		return nil, err
-	}
-
-	candidates := []*NetworkWithExternalExt{}
-	for i, n := range existingNetworks {
-		if n.External == isExternal {
-			candidates = append(candidates, &existingNetworks[i])
-		}
-	}
-
-	switch len(candidates) {
-	case 1:
-		return candidates[0], nil
-	case 0:
-		return nil, fmt.Errorf("network named '%s' with external=%v not found", name, isExternal)
-	default:
-		return nil, fmt.Errorf("found %d networks for name '%s' (external=%v), expected exactly one", len(candidates), name, isExternal)
-	}
-}
-
-func getDefaultExternalNetwork(netClient *gophercloud.ServiceClient) (*NetworkWithExternalExt, error) {
-	existingNetworks, err := getAllNetworks(netClient, osnetworks.ListOpts{})
-	if err != nil {
-		return nil, err
-	}
-
-	for _, n := range existingNetworks {
-		if n.External {
-			return &n, nil
-		}
-	}
-
-	return nil, errors.New("no external network found")
-}
-
-func validateSecurityGroupsExist(netClient *gophercloud.ServiceClient, securityGroups []string) error {
-	for _, sg := range securityGroups {
-		results, err := getSecurityGroups(netClient, ossecuritygroups.ListOpts{Name: sg})
-		if err != nil {
-			return fmt.Errorf("failed to get security group: %w", err)
-		}
-		if len(results) == 0 {
-			return fmt.Errorf("specified security group %s not found", sg)
-		}
-	}
-	return nil
-}
-
-func deleteSecurityGroup(netClient *gophercloud.ServiceClient, sgName string) error {
-	results, err := getSecurityGroups(netClient, ossecuritygroups.ListOpts{Name: sgName})
-	if err != nil {
-		return fmt.Errorf("failed to get security group: %w", err)
-	}
-
-	for _, sg := range results {
-		res := ossecuritygroups.Delete(netClient, sg.ID)
-		if res.Err != nil {
-			return res.Err
-		}
-		if err := res.ExtractErr(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-type createSecurityGroupRequest struct {
-	secGroupName   string
-	ipv4Rules      bool
-	ipv6Rules      bool
-	nodePortsCIDRs kubermaticv1.NetworkRanges
-	lowPort        int
-	highPort       int
-}
-
-func createSecurityGroup(netClient *gophercloud.ServiceClient, req createSecurityGroupRequest) (string, error) {
-	secGroups, err := getSecurityGroups(netClient, ossecuritygroups.ListOpts{Name: req.secGroupName})
-	if err != nil {
-		return "", fmt.Errorf("failed to get security groups: %w", err)
-	}
-
-	var securityGroupID string
-	switch len(secGroups) {
-	case 0:
-		gres := ossecuritygroups.Create(netClient, ossecuritygroups.CreateOpts{
-			Name:        req.secGroupName,
-			Description: "Contains security rules for the Kubernetes worker nodes",
-		})
-		if gres.Err != nil {
-			return "", gres.Err
-		}
-		g, err := gres.Extract()
-		if err != nil {
-			return "", err
-		}
-		securityGroupID = g.ID
-	case 1:
-		securityGroupID = secGroups[0].ID
-	default:
-		return "", fmt.Errorf("there are already %d security groups with name %q, dont know which one to use",
-			len(secGroups), req.secGroupName)
-	}
-
-	var rules []ossecuritygrouprules.CreateOpts
-
-	if req.ipv4Rules {
-		rules = append(rules, []ossecuritygrouprules.CreateOpts{
-			{
-				// Allows ipv4 traffic within this group
-				Direction:     ossecuritygrouprules.DirIngress,
-				EtherType:     ossecuritygrouprules.EtherType4,
-				SecGroupID:    securityGroupID,
-				RemoteGroupID: securityGroupID,
-			},
-			{
-				// Allows ssh from external
-				Direction:    ossecuritygrouprules.DirIngress,
-				EtherType:    ossecuritygrouprules.EtherType4,
-				SecGroupID:   securityGroupID,
-				PortRangeMin: provider.DefaultSSHPort,
-				PortRangeMax: provider.DefaultSSHPort,
-				Protocol:     ossecuritygrouprules.ProtocolTCP,
-			},
-			{
-				// Allows ICMP traffic
-				Direction:  ossecuritygrouprules.DirIngress,
-				EtherType:  ossecuritygrouprules.EtherType4,
-				SecGroupID: securityGroupID,
-				Protocol:   ossecuritygrouprules.ProtocolICMP,
-			},
-		}...)
-	}
-
-	if req.ipv6Rules {
-		rules = append(rules, []ossecuritygrouprules.CreateOpts{
-			{
-				// Allows ipv6 traffic within this group
-				Direction:     ossecuritygrouprules.DirIngress,
-				EtherType:     ossecuritygrouprules.EtherType6,
-				SecGroupID:    securityGroupID,
-				RemoteGroupID: securityGroupID,
-			},
-			{
-				// Allows ssh from external
-				Direction:    ossecuritygrouprules.DirIngress,
-				EtherType:    ossecuritygrouprules.EtherType6,
-				SecGroupID:   securityGroupID,
-				PortRangeMin: provider.DefaultSSHPort,
-				PortRangeMax: provider.DefaultSSHPort,
-				Protocol:     ossecuritygrouprules.ProtocolTCP,
-			},
-			{
-				// Allows ICMPv6 traffic
-				Direction:  ossecuritygrouprules.DirIngress,
-				EtherType:  ossecuritygrouprules.EtherType6,
-				SecGroupID: securityGroupID,
-				Protocol:   ossecuritygrouprules.ProtocolIPv6ICMP,
-			},
-		}...)
-	}
-
-	for _, cidr := range req.nodePortsCIDRs.CIDRBlocks {
-		tcp := ossecuritygrouprules.CreateOpts{
-			// Allows TCP traffic to nodePorts from external
-			Direction:      ossecuritygrouprules.DirIngress,
-			SecGroupID:     securityGroupID,
-			PortRangeMin:   req.lowPort,
-			PortRangeMax:   req.highPort,
-			Protocol:       ossecuritygrouprules.ProtocolTCP,
-			RemoteIPPrefix: cidr,
-		}
-		udp := ossecuritygrouprules.CreateOpts{
-			// Allows UDP traffic to nodePorts from external
-			Direction:      ossecuritygrouprules.DirIngress,
-			SecGroupID:     securityGroupID,
-			PortRangeMin:   req.lowPort,
-			PortRangeMax:   req.highPort,
-			Protocol:       ossecuritygrouprules.ProtocolUDP,
-			RemoteIPPrefix: cidr,
-		}
-		if net.IsIPv4CIDRString(cidr) {
-			tcp.EtherType = ossecuritygrouprules.EtherType4
-			udp.EtherType = ossecuritygrouprules.EtherType4
-		} else {
-			tcp.EtherType = ossecuritygrouprules.EtherType6
-			udp.EtherType = ossecuritygrouprules.EtherType6
-		}
-		rules = append(rules, tcp, udp)
-	}
-
-	for _, opts := range rules {
-	reiterate:
-		rres := ossecuritygrouprules.Create(netClient, opts)
-		if rres.Err != nil {
-			var unexpected gophercloud.ErrUnexpectedResponseCode
-			if errors.As(rres.Err, &unexpected) && unexpected.Actual == http.StatusConflict {
-				// already exists
-				continue
-			}
-
-			if errors.As(rres.Err, &gophercloud.ErrDefault400{}) && opts.Protocol == ossecuritygrouprules.ProtocolIPv6ICMP {
-				// workaround for old versions of Openstack with different protocol name,
-				// from before https://review.opendev.org/#/c/252155/
-				opts.Protocol = "icmpv6"
-				goto reiterate // I'm very sorry, but this was really the cleanest way.
-			}
-
-			return "", rres.Err
-		}
-
-		if _, err := rres.Extract(); err != nil {
-			return "", err
-		}
-	}
-
-	return req.secGroupName, nil
-}
-
-func createUserClusterNetwork(netClient *gophercloud.ServiceClient, clusterName string) (*osnetworks.Network, error) {
-	iTrue := true
-	res := osnetworks.Create(netClient, osnetworks.CreateOpts{
-		Name:         resourceNamePrefix + clusterName,
-		AdminStateUp: &iTrue,
-	})
-	if res.Err != nil {
-		return nil, res.Err
-	}
-	return res.Extract()
-}
-
-func deleteNetworkByName(netClient *gophercloud.ServiceClient, networkName string) error {
-	network, err := getNetworkByName(netClient, networkName, false)
-	if err != nil {
-		return fmt.Errorf("failed to get network '%s' by name: %w", networkName, err)
-	}
-
-	res := osnetworks.Delete(netClient, network.ID)
-	if res.Err != nil {
-		return res.Err
-	}
-	return res.ExtractErr()
-}
-
-func deleteSubnet(netClient *gophercloud.ServiceClient, subnetID string) error {
-	res := ossubnets.Delete(netClient, subnetID)
-	if res.Err != nil {
-		return res.Err
-	}
-	return res.ExtractErr()
-}
-
-func deleteRouter(netClient *gophercloud.ServiceClient, routerID string) error {
-	res := osrouters.Delete(netClient, routerID)
-	if res.Err != nil {
-		return res.Err
-	}
-	return res.ExtractErr()
-}
-
-func createSubnet(netClient *gophercloud.ServiceClient, clusterName, networkID string, dnsServers []string) (*ossubnets.Subnet, error) {
-	iTrue := true
-	subnetOpts := ossubnets.CreateOpts{
-		Name:       resourceNamePrefix + clusterName,
-		NetworkID:  networkID,
-		IPVersion:  gophercloud.IPv4,
-		CIDR:       subnetCIDR,
-		GatewayIP:  nil,
-		EnableDHCP: &iTrue,
-		AllocationPools: []ossubnets.AllocationPool{
-			{
-				Start: subnetFirstAddress,
-				End:   subnetLastAddress,
-			},
-		},
-	}
-
-	for _, s := range dnsServers {
-		if net.IsIPv4String(s) {
-			subnetOpts.DNSNameservers = append(subnetOpts.DNSNameservers, s)
-		}
-	}
-
-	res := ossubnets.Create(netClient, subnetOpts)
-	if res.Err != nil {
-		return nil, res.Err
-	}
-	return res.Extract()
-}
-
-func createIPv6Subnet(netClient *gophercloud.ServiceClient, clusterName, networkID, subnetPoolName string, dnsServers []string) (*ossubnets.Subnet, error) {
-	subnetOpts := ossubnets.CreateOpts{
-		Name:            resourceNamePrefix + clusterName + "-ipv6",
-		NetworkID:       networkID,
-		IPVersion:       gophercloud.IPv6,
-		GatewayIP:       nil,
-		EnableDHCP:      ptr.To(true),
-		IPv6AddressMode: "dhcpv6-stateless",
-		IPv6RAMode:      "dhcpv6-stateless",
-	}
-	subnetPoolID := ""
-
-	// if IPv6 subnet pool name is provided - resolve to ID
-	if subnetPoolName != "" {
-		subnetPool, err := getSubnetPoolByName(netClient, subnetPoolName)
-		if err != nil {
-			return nil, err
-		}
-		subnetPoolID = subnetPool.ID
-	}
-
-	// if IPv6 subnet pool name is not provided - look for the default IPv6 subnet pool
-	if subnetPoolID == "" {
-		pools, err := getAllSubnetPools(netClient, subnetpools.ListOpts{IPVersion: 6, IsDefault: ptr.To(true)})
-		if err != nil {
-			return nil, err
-		}
-		if len(pools) > 0 {
-			subnetPoolID = pools[0].ID
-		}
-	}
-
-	if subnetPoolID != "" {
-		subnetOpts.SubnetPoolID = subnetPoolID
-	} else {
-		// if no IPv6 subnet pool was provided / found, use the default IPv6 subnet CIDR
-		subnetOpts.CIDR = defaultIPv6SubnetCIDR
-	}
-
-	for _, s := range dnsServers {
-		if net.IsIPv6String(s) {
-			subnetOpts.DNSNameservers = append(subnetOpts.DNSNameservers, s)
-		}
-	}
-
-	res := ossubnets.Create(netClient, subnetOpts)
-	if res.Err != nil {
-		return nil, res.Err
-	}
-	return res.Extract()
-}
-
-func getSubnetPoolByName(netClient *gophercloud.ServiceClient, name string) (*subnetpools.SubnetPool, error) {
-	pools, err := getAllSubnetPools(netClient, subnetpools.ListOpts{Name: name})
-	if err != nil {
-		return nil, err
-	}
-	switch len(pools) {
-	case 1:
-		return &pools[0], nil
-	case 0:
-		return nil, fmt.Errorf("subnet pool named '%s' not found", name)
-	default:
-		return nil, fmt.Errorf("found %d subnet pools for name '%s', expected exactly one", len(pools), name)
-	}
-}
-
-func getSubnetByID(netClient *gophercloud.ServiceClient, id string) (*ossubnets.Subnet, error) {
-	listOpts := ossubnets.ListOpts{
-		ID: id,
-	}
-	allSubnets, err := getAllSubnets(netClient, listOpts)
-	if err != nil {
-		return nil, err
-	}
-	switch len(allSubnets) {
-	case 1:
-		return &allSubnets[0], nil
-	case 0:
-		return nil, fmt.Errorf("subnet with id '%s' not found", id)
-	default:
-		return nil, fmt.Errorf("found %d subnets for id '%s', expected exactly one", len(allSubnets), id)
-	}
-}
-
-func getSubnetByName(netClient *gophercloud.ServiceClient, name string) (*ossubnets.Subnet, error) {
-	listOpts := ossubnets.ListOpts{
-		Name: name,
-	}
-	allSubnets, err := getAllSubnets(netClient, listOpts)
-	if err != nil {
-		return nil, err
-	}
-	switch len(allSubnets) {
-	case 1:
-		return &allSubnets[0], nil
-	case 0:
-		return nil, fmt.Errorf("subnet named '%s' not found", name)
-	default:
-		return nil, fmt.Errorf("found %d subnets for name '%s', expected exactly one", len(allSubnets), name)
-	}
-}
-
-func getAllSubnets(netClient *gophercloud.ServiceClient, listOpts ossubnets.ListOpts) ([]ossubnets.Subnet, error) {
-	allPages, err := ossubnets.List(netClient, listOpts).AllPages()
-	if err != nil {
-		return nil, err
-	}
-	allSubnets, err := ossubnets.ExtractSubnets(allPages)
-	if err != nil {
-		return nil, err
-	}
-	return allSubnets, nil
-}
-
-func getAllSubnetPools(netClient *gophercloud.ServiceClient, listOpts subnetpools.ListOpts) ([]subnetpools.SubnetPool, error) {
-	allPages, err := subnetpools.List(netClient, listOpts).AllPages()
-	if err != nil {
-		return nil, err
-	}
-	allSubnetPools, err := subnetpools.ExtractSubnetPools(allPages)
-	if err != nil {
-		return nil, err
-	}
-	return allSubnetPools, nil
-}
-
-func createRouter(netClient *gophercloud.ServiceClient, clusterName, extNetworkName string) (*osrouters.Router, error) {
-	extNetwork, err := getNetworkByName(netClient, extNetworkName, true)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get external network %q: %w", extNetworkName, err)
-	}
-
-	iTrue := true
-	gwi := osrouters.GatewayInfo{
-		NetworkID: extNetwork.ID,
-	}
-
-	res := osrouters.Create(netClient, osrouters.CreateOpts{
-		Name:         resourceNamePrefix + clusterName,
-		AdminStateUp: &iTrue,
-		GatewayInfo:  &gwi,
-	})
-	if res.Err != nil {
-		return nil, res.Err
-	}
-	return res.Extract()
-}
-
-func attachSubnetToRouter(netClient *gophercloud.ServiceClient, subnetID, routerID string) (*osrouters.InterfaceInfo, error) {
-	interf, err := func() (*osrouters.InterfaceInfo, error) {
-		res := osrouters.AddInterface(netClient, routerID, osrouters.AddInterfaceOpts{
-			SubnetID: subnetID,
-		})
-		if res.Err != nil {
-			return nil, res.Err
-		}
-		return res.Extract()
-	}()
-	return interf, ignoreRouterAlreadyHasPortInSubnetError(err, subnetID)
-}
-
-func detachSubnetFromRouter(netClient *gophercloud.ServiceClient, subnetID, routerID string) (*osrouters.InterfaceInfo, error) {
-	res := osrouters.RemoveInterface(netClient, routerID, osrouters.RemoveInterfaceOpts{
-		SubnetID: subnetID,
-	})
-	if res.Err != nil {
-		return nil, res.Err
-	}
-	return res.Extract()
-}
 
 func getFlavors(authClient *gophercloud.ProviderClient, region string) ([]osflavors.Flavor, error) {
 	computeClient, err := goopenstack.NewComputeV2(authClient, gophercloud.EndpointOpts{Availability: gophercloud.AvailabilityPublic, Region: region})
@@ -661,33 +106,6 @@ func getTenants(authClient *gophercloud.ProviderClient, region string) ([]osproj
 	return allProjects, nil
 }
 
-func getSubnetForNetwork(netClient *gophercloud.ServiceClient, networkIDOrName string) ([]ossubnets.Subnet, error) {
-	var allSubnets []ossubnets.Subnet
-
-	networks, err := getAllNetworks(netClient, osnetworks.ListOpts{Name: networkIDOrName})
-	if err != nil {
-		return nil, fmt.Errorf("failed to list networks: %w", err)
-	}
-
-	networkID := networkIDOrName
-	if len(networks) == 1 {
-		networkID = networks[0].ID
-	} else if len(networks) > 1 {
-		return nil, fmt.Errorf("got %d networks for idOrName '%s', expected one at most", len(networks), networkIDOrName)
-	}
-
-	allPages, err := ossubnets.List(netClient, ossubnets.ListOpts{NetworkID: networkID}).AllPages()
-	if err != nil {
-		return nil, err
-	}
-
-	if allSubnets, err = ossubnets.ExtractSubnets(allPages); err != nil {
-		return nil, err
-	}
-
-	return allSubnets, nil
-}
-
 func isNotFoundErr(err error) bool {
 	var errNotFound gophercloud.ErrDefault404
 
@@ -699,38 +117,6 @@ func isEndpointNotFoundErr(err error) bool {
 	// left side of the || to catch any error returned as pointer to struct (current case of gophercloud)
 	// right side of the || to catch any error returned as struct (in case...)
 	return errors.As(err, &endpointNotFoundErr) || errors.As(err, &gophercloud.ErrEndpointNotFound{})
-}
-
-func getRouterIDForSubnet(netClient *gophercloud.ServiceClient, subnetID string) (string, error) {
-	ports, err := getAllNetworkPorts(netClient, subnetID)
-	if err != nil {
-		return "", fmt.Errorf("failed to list ports for subnet: %w", err)
-	}
-
-	for _, port := range ports {
-		if port.DeviceOwner == "network:router_interface" || port.DeviceOwner == "network:router_interface_distributed" || port.DeviceOwner == "network:ha_router_replicated_interface" {
-			// Check IP for the interface & check if the IP belongs to the subnet
-			return port.DeviceID, nil
-		}
-	}
-
-	return "", nil
-}
-
-func getAllNetworkPorts(netClient *gophercloud.ServiceClient, subnetID string) ([]osports.Port, error) {
-	allPages, err := osports.List(netClient, osports.ListOpts{
-		FixedIPs: []osports.FixedIPOpts{{SubnetID: subnetID}},
-	}).AllPages()
-	if err != nil {
-		return nil, err
-	}
-
-	allPorts, err := osports.ExtractPorts(allPages)
-	if err != nil {
-		return nil, err
-	}
-
-	return allPorts, nil
 }
 
 func getAvailabilityZones(computeClient *gophercloud.ServiceClient) ([]osavailabilityzones.AvailabilityZone, error) {
